Move benchmarks into convert tests
